requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.11.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0.1
pydantic>=2.5.0
